src/requirements.txt
# Use these dependencies for local development
# including all scripts for provisioning and deploying

# openai SDK
openai==1.13.3

# promptflow packages
promptflow[azure]==1.10.1
promptflow-tracing==1.10.1
promptflow-tools==1.4.0
promptflow-evals==0.2.0.dev0

# azure dependencies
azure-core==1.30.1
azure-identity==1.16.0
azure-mgmt-resource==23.0.1
azure-mgmt-search==9.1.0
azure-mgmt-cognitiveservices==13.5.0
azure-ai-ml==1.16.0

# utilities
omegaconf-argparse==1.0.1
omegaconf==2.3.0
pydantic>=2.6
